refactor: separate out more concerns
